requirements1.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
passlib
sqlalchemy
pytest
bcrypt
python-jose